requirements.txt
asgiref==3.7.2
astroid==2.15.5
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
geographiclib==2.0
geopy==2.4.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
platformdirs==3.8.0
psycopg==3.1.10
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyrsistent==0.19.3
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0
requests==2.31.0
redgreenunittest==0.1.1
simplejson==3.19.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.2
uritemplate==4.1.1
# urllib3 version changed from 2.0.3 to 1.26.16
# inorder to satisfy the requirement of the dependency
# botocore 1.31.53 which requires urllib3<1.27,>=1.25.4
urllib3==1.26.16 
wrapt==1.15.0
drf-nested-routers==0.93.4
boto3==1.28.52 
django-storages==1.14    
django-environ==0.11.2  